Extract spec orchestration into mapper index and context modules
Move the flatten/emit/collect orchestration out of mapApp into a new
convertWaspSpecToAppSpec entry point (index.ts), and the context builder
plus makeRefParser into context.ts. mapApp now maps only the App decl.
Wire appAnalyzer and tests to convertWaspSpecToAppSpec.
