+++ waspc/data/packages/spec/__tests__/spec/mapApp.unit.test.ts
@@ -2,8 +2,12 @@ import { describe, expect, test } from "vitest";
 import * as AppSpec from "../../src/appSpec.js";
 import { normalizePrerender } from "../../src/normalizePrerender.js";
 import * as AppSpecMapper from "../../src/spec/mapper/app.js";
+import {
+  type AppMapperContext,
+  makeRefParser,
+} from "../../src/spec/mapper/context.js";
+import { convertWaspSpecToAppSpec } from "../../src/spec/mapper/index.js";
 import * as SpecElementMapper from "../../src/spec/mapper/specElements.js";
-import type { AppMapperContext } from "../../src/spec/mapper/types.js";
 import {
   api,
   app,
@@ -30,10 +34,10 @@ function makeMapperContext({
   routeNames?: string[];
 } = {}): AppMapperContext {
   const ctx: AppMapperContext = {
-    emitEntityRef: AppSpecMapper.makeRefParser("Entity", entityNames),
-    emitRouteRef: AppSpecMapper.makeRefParser("Route", routeNames),
+    emitEntityRef: makeRefParser("Entity", entityNames),
+    emitRouteRef: makeRefParser("Route", routeNames),
     emitRefObject: mapRefObjectForMockProjectDir,
-    // Decl registration (dedup and conflict detection) is a `mapApp`-level
+    // Decl registration (dedup and conflict detection) is an orchestration-level
     // concern; tests exercising individual mappers only need the ref.
     emitSpecElementRef<SpecElement extends WaspSpec.SpecElement>(
       specElement: SpecElement,
@@ -53,13 +57,13 @@ function getSpecElementDeclName(specElement: WaspSpec.SpecElement): string {
 }
 
 function mapMockApp(app: WaspSpec.App, entityNames: string[]) {
-  return AppSpecMapper.mapApp(app, {
+  return convertWaspSpecToAppSpec(app, {
     entityNames,
     projectRootDir: Fixtures.MOCK_PROJECT_DIR,
   });
 }
 
-describe("mapApp", () => {
+describe("convertWaspSpecToAppSpec", () => {
   test("should map minimal app correctly", () => {
     const entityNames = Fixtures.getEntities("minimal");
     const app = Fixtures.getApp("minimal");
